extracted implementation part of `Type.h` to `Type.c`
